Back to SW DDS configuration
Went back to DDS software configuration since using SPORTs was overkill
and generated problems.
Synchronization of the DDS is accomplished through a very stable initial
configuration. As such, changing SRUs (such as the approach when using
the SPORTs, when the DATA line stayed the same but the W_CLK_# changed)
led to sync failure of the DDS.
Also, the SPORTs method required that the processor would hang while
waiting for the SPORT transfer to be completed. Since this is an
external event, a SPORTs transfer that did not complete, would hang the
processor indefinitely -> Bad approach.
Bit banging is the way to go! Also, programming speeds are acceptable.
Around 41us at the moment. But can be easily reduced.